Add files via upload
Image classification is the task of assigning a label to an image based on its content. In this project, we will use TensorFlow, a powerful open-source deep learning library, to build a machine learning model that can classify images. We'll also use Streamlit, a Python framework for creating interactive web applications, to deploy the model and allow users to interact with it in real-time.